refactor(telemetry) relocating utilities into the @utils sub-module
